+++ src/contents.rs
@@ -32,7 +32,7 @@
 
 use crate::{Buffer, ContentsRef};
 use num::PrimInt;
-use std::borrow::{Borrow, BorrowMut};
+use std::borrow::Borrow;
 use std::mem;
 use std::ops::{Deref, DerefMut};
 
@@ -190,7 +190,7 @@ impl Contents {
     /// let bytes: &[u8] = &[1, 2, 3];
     /// let contents = Contents::from_bytes(bytes);
     ///
-    /// assert_eq!(contents.as_bytes(), bytes);
+    /// assert_eq!(contents.as_ref() as &[u8], bytes);
     /// ```
     pub fn from_bytes(bytes: &[u8]) -> Self {
         Self {
@@ -302,7 +302,7 @@ impl Contents {
 
 impl AsRef<[u8]> for Contents {
     fn as_ref(&self) -> &[u8] {
-        self.as_bytes()
+        self.deref().as_ref()
     }
 }
 
@@ -314,7 +314,7 @@ impl AsRef<ContentsRef> for Contents {
 
 impl AsMut<[u8]> for Contents {
     fn as_mut(&mut self) -> &mut [u8] {
-        self.as_mut_bytes()
+        self.deref_mut().as_mut()
     }
 }
 
@@ -326,12 +326,6 @@ impl AsMut<ContentsRef> for Contents {
 
 impl Borrow<ContentsRef> for Contents {
     fn borrow(&self) -> &ContentsRef {
-        self
-    }
-}
-
-impl BorrowMut<ContentsRef> for Contents {
-    fn borrow_mut(&mut self) -> &mut ContentsRef {
         self
     }
 }
@@ -369,7 +363,7 @@ mod tests {
             let contents = Contents::from_integer(i);
             let expected: &[u8] = &[i as u8];
 
-            assert_eq!(contents.as_bytes(), expected);
+            assert_eq!(contents.as_ref() as &[u8], expected);
         }
     }
 
@@ -378,13 +372,13 @@ mod tests {
         for i in 0..0x80 {
             let contents = Contents::from_integer(i as u8);
             let expected: &[u8] = &[i as u8];
-            assert_eq!(contents.as_bytes(), expected);
+            assert_eq!(contents.as_ref() as &[u8], expected);
         }
 
         for i in 0x80..=u8::MAX {
             let contents = Contents::from_integer(i as u8);
             let expected: &[u8] = &[0x00, i];
-            assert_eq!(contents.as_bytes(), expected);
+            assert_eq!(contents.as_ref() as &[u8], expected);
         }
     }
 
@@ -399,7 +393,7 @@ mod tests {
             let f = i.unsigned_shr(8) as u8;
             let s = i as u8;
             let expected = &[f, s];
-            assert_eq!(contents.as_bytes(), expected);
+            assert_eq!(contents.as_ref() as &[u8], expected);
         }
     }
 
@@ -412,7 +406,7 @@ mod tests {
             let s = i as u8;
             let expected: &[u8] = &[0, f, s];
 
-            assert_eq!(contents.as_bytes(), expected);
+            assert_eq!(contents.as_ref() as &[u8], expected);
         }
     }
 
@@ -425,7 +419,7 @@ mod tests {
             let mut expected: [u8; 16] = [0x00; 16];
             expected[0] = 0x80;
 
-            assert_eq!(contents.as_bytes(), expected);
+            assert_eq!(contents.as_ref() as &[u8], expected);
         }
 
         // i128::MAX
@@ -435,7 +429,7 @@ mod tests {
             let mut expected: [u8; 16] = [0xff; 16];
             expected[0] = 0x7f;
 
-            assert_eq!(contents.as_bytes(), expected);
+            assert_eq!(contents.as_ref() as &[u8], expected);
         }
     }
 
@@ -448,7 +442,7 @@ mod tests {
             let mut expected: [u8; 17] = [0x00; 17];
             expected[1] = 0x80;
 
-            assert_eq!(contents.as_bytes(), expected);
+            assert_eq!(contents.as_ref() as &[u8], expected);
         }
 
         // u128::MAX
@@ -458,7 +452,7 @@ mod tests {
             let mut expected: [u8; 17] = [0xff; 17];
             expected[0] = 0x00;
 
-            assert_eq!(contents.as_bytes(), expected);
+            assert_eq!(contents.as_ref() as &[u8], expected);
         }
     }
 
@@ -878,13 +872,13 @@ mod tests {
         // True
         {
             let contents = Contents::from(true);
-            assert_eq!(&[0xff], contents.as_bytes());
+            assert_eq!(&[0xff], contents.as_ref() as &[u8]);
         }
 
         // false
         {
             let contents = Contents::from(false);
-            assert_eq!(&[0x00], contents.as_bytes());
+            assert_eq!(&[0x00], contents.as_ref() as &[u8]);
         }
     }
 }