+++ src/identifier_ref.rs
@@ -44,12 +44,9 @@ pub const PC_MASK: u8 = 0x20;
 
 /// `IdRef` is a wrapper of `[u8]` representing Identifier.
 ///
-/// The user can access the inner slice via method [`as_bytes`] or [`as_mut_bytes`]
+/// The user can access the inner slice via `AsRef` implementation.
 ///
 /// This struct is `Unsized`, and the user will usually use a reference to it.
-///
-/// [`as_bytes`]: Self::as_bytes
-/// [`as_mut_bytes`]: Self::as_mut_bytes
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct IdRef {
     bytes: [u8],
@@ -1067,7 +1064,7 @@ impl ToOwned for IdRef {
     type Owned = Id;
 
     fn to_owned(&self) -> Self::Owned {
-        unsafe { Self::Owned::from_bytes_unchecked(self.as_bytes()) }
+        unsafe { Self::Owned::from_bytes_unchecked(self.as_ref()) }
     }
 }
 
@@ -1105,7 +1102,7 @@ impl IdRef {
     /// assert_eq!(id.len(), bytes.len());
     /// ```
     pub fn len(&self) -> usize {
-        return self.as_bytes().len();
+        return self.as_ref().len();
     }
 
     /// Returns the `ClassTag` of `self`.
@@ -1276,9 +1273,9 @@ impl IdRef {
             debug_assert_eq!(self.bytes[0] & LONG_FLAG, LONG_FLAG);
 
             let mask: u8 = !MORE_FLAG;
-            let mut ret = T::from_u8(self.as_bytes()[1] & mask).unwrap();
+            let mut ret = T::from_u8(self.as_ref()[1] & mask).unwrap();
 
-            for &octet in self.as_bytes()[2..].iter() {
+            for &octet in self.as_ref()[2..].iter() {
                 let shft_mul = T::from_u8(128).ok_or(Error::OverFlow)?;
                 ret = ret.checked_mul(&shft_mul).ok_or(Error::OverFlow)?;
                 ret = ret + T::from_u8(octet & mask).unwrap();
@@ -1286,49 +1283,6 @@ impl IdRef {
 
             Ok(ret)
         }
-    }
-
-    /// Provides a reference to the inner slice.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bsn1::IdRef;
-    ///
-    /// let bytes: &[u8] = &[3];
-    /// let idref = unsafe { IdRef::from_bytes_unchecked(bytes) };
-    /// assert_eq!(bytes, idref.as_bytes());
-    /// ```
-    pub fn as_bytes(&self) -> &[u8] {
-        &self.bytes
-    }
-
-    /// Provides a mutable reference to the inner slice.
-    ///
-    /// # Safety
-    ///
-    /// The behaviour is undefined if the inner slice will be invalid as an Identifier of 'ASN.1'.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bsn1::IdRef;
-    ///
-    /// let bytes: &mut [u8] = &mut [5];
-    ///
-    /// unsafe {
-    ///     let idref = unsafe { IdRef::from_mut_bytes_unchecked(bytes) };
-    ///     assert_eq!(5, idref.as_bytes()[0]);
-    ///
-    ///     idref.as_mut_bytes()[0] = 6;
-    ///     assert_eq!(6, idref.as_bytes()[0]);
-    /// }
-    ///
-    /// // 'bytes' is updated as well.
-    /// assert_eq!(6, bytes[0]);
-    /// ```
-    pub unsafe fn as_mut_bytes(&mut self) -> &mut [u8] {
-        &mut self.bytes
     }
 
     /// Update the class tag of `self`.
@@ -1339,7 +1293,7 @@ impl IdRef {
     /// use bsn1::{ClassTag, IdRef};
     ///
     /// // Creates a '&mut IdRef' representing 'Universal Integer'.
-    /// let mut bytes = Vec::from(IdRef::integer().as_bytes());
+    /// let mut bytes = Vec::from(IdRef::integer().as_ref());
     /// let idref = IdRef::parse_mut(&mut bytes).unwrap();
     ///
     /// assert_eq!(ClassTag::Universal, idref.class());
@@ -1348,7 +1302,7 @@ impl IdRef {
     /// assert_eq!(ClassTag::Application, idref.class());
     /// ```
     pub fn set_class(&mut self, cls: ClassTag) {
-        let bytes = unsafe { self.as_mut_bytes() };
+        let bytes = &mut self.bytes;
 
         bytes[0] &= !CLASS_MASK;
         bytes[0] |= cls as u8;
@@ -1362,7 +1316,7 @@ impl IdRef {
     /// use bsn1::{PCTag, IdRef};
     ///
     /// // Creates a '&mut IdRef' representing 'Universal Integer'.
-    /// let mut bytes = Vec::from(IdRef::integer().as_bytes());
+    /// let mut bytes = Vec::from(IdRef::integer().as_ref());
     /// let idref = IdRef::parse_mut(&mut bytes).unwrap();
     ///
     /// assert_eq!(PCTag::Primitive, idref.pc());
@@ -1371,7 +1325,7 @@ impl IdRef {
     /// assert_eq!(PCTag::Constructed, idref.pc());
     /// ```
     pub fn set_pc(&mut self, pc: PCTag) {
-        let bytes = unsafe { self.as_mut_bytes() };
+        let bytes = &mut self.bytes;
 
         const MASK: u8 = 0xdf;
         bytes[0] &= MASK;
@@ -1400,12 +1354,12 @@ mod tests {
                     let first = cl as u8 + pc as u8 + 0;
                     let bytes: &[u8] = &[first];
                     let id = IdRef::parse(bytes).unwrap();
-                    assert_eq!(bytes, id.as_bytes());
+                    assert_eq!(bytes, id.as_ref());
 
                     let first = cl as u8 + pc as u8 + 0x1e;
                     let bytes: &[u8] = &[first];
                     let id = IdRef::parse(bytes).unwrap();
-                    assert_eq!(bytes, id.as_bytes());
+                    assert_eq!(bytes, id.as_ref());
                 }
 
                 let first = cl as u8 + pc as u8 + 0x1f;
@@ -1414,11 +1368,11 @@ mod tests {
                 {
                     let bytes: &[u8] = &[first, 0x1f];
                     let id = IdRef::parse(bytes).unwrap();
-                    assert_eq!(bytes, id.as_bytes());
+                    assert_eq!(bytes, id.as_ref());
 
                     let bytes: &[u8] = &[first, 0x7f];
                     let id = IdRef::parse(bytes).unwrap();
-                    assert_eq!(bytes, id.as_bytes());
+                    assert_eq!(bytes, id.as_ref());
                 }
 
                 // len bytes
@@ -1432,7 +1386,7 @@ mod tests {
 
                     let bytes: &[u8] = &bytes;
                     let id = IdRef::parse(bytes).unwrap();
-                    assert_eq!(bytes, id.as_bytes());
+                    assert_eq!(bytes, id.as_ref());
 
                     let mut bytes = vec![first];
                     for _ in 1..len {
@@ -1442,7 +1396,7 @@ mod tests {
 
                     let bytes: &[u8] = &bytes;
                     let id = IdRef::parse(bytes).unwrap();
-                    assert_eq!(bytes, id.as_bytes());
+                    assert_eq!(bytes, id.as_ref());
                 }
             }
         }